Reformat most files + improve signout through activeAccessToken flag
Change (") to (') in all cases + add an 'activeAccessToken' field to User that is set to 'true' when the User is signed in, and set to 'false' when the User is signed out. This way, we can immediately discard the access token when we sign out, without waiting for it to discard itself.